tests: split AssignmentRoutesTests.swift (2,394 lines → 6 files)
Follow-on to #490.  The mega-class held 43 tests and was a single
worker thread under --parallel.  Splitting along the existing MARK
boundaries gives XCTest five concrete subclasses to schedule across
workers, with shared setUp/tearDown/fixtures lifted to a common base.

Layout:
  AssignmentRoutesTestCase.swift       — base class, helpers, fixtures
  AssignmentRoutesDashboardTests.swift — GET /instructor + close-expired (8)
  AssignmentRoutesPublishTests.swift   — POST /instructor + create regressions (18)
  AssignmentRoutesLifecycleTests.swift — open/close/delete simple paths (5)
  AssignmentRoutesRetestTests.swift    — retest single + all + suite metadata (7)
  AssignmentRoutesNotebookTests.swift  — reset-notebook + course student subs (5)

Total: 43 tests preserved (8+18+5+7+5).  Behaviour-preserving refactor;
all tests pass under --parallel.
