+++ Tests/APITests/AssignmentRoutesLifecycleTests.swift
@@ -1,0 +1,120 @@
+// Tests/APITests/AssignmentRoutesLifecycleTests.swift
+//
+// Split from AssignmentRoutesTests.swift.  See AssignmentRoutesTestCase.swift
+// for shared helpers (auth, fixtures, multipart builders, zip + notebook
+// staging).
+
+import Core
+import Fluent
+import Foundation
+import XCTVapor
+import XCTest
+
+@testable import chickadee_server
+
+final class AssignmentRoutesLifecycleTests: AssignmentRoutesTestCase {
+
+    // MARK: - POST /instructor/:id/open
+
+    func testOpenAssignmentSetsIsOpenTrue() async throws {
+        let cookie = try await loginAsInstructor()
+        let (csrf, sessionCookie) = try await csrfFields(for: "/instructor", cookie: cookie, on: app)
+        try await insertSetup(id: "setup_open")
+        let a = try await insertAssignment(testSetupID: "setup_open", title: "Draft", isOpen: false)
+        let id = a.publicID
+
+        try await app.asyncTest(
+            .POST, "/instructor/\(id)/open",
+            beforeRequest: { req in
+                req.headers.add(name: .cookie, value: sessionCookie)
+                try req.content.encode(["_csrf": csrf], as: .urlEncodedForm)
+            },
+            afterResponse: { res in
+                XCTAssertEqual(res.status, .seeOther)
+                XCTAssertEqual(res.headers.first(name: .location), "/instructor")
+            })
+
+        let updated = try await APIAssignment.find(a.id, on: app.db)
+        XCTAssertEqual(updated?.isOpen, true)
+    }
+
+    // MARK: - POST /instructor/:id/close
+
+    func testCloseAssignmentSetsIsOpenFalse() async throws {
+        let cookie = try await loginAsInstructor()
+        let (csrf, sessionCookie) = try await csrfFields(for: "/instructor", cookie: cookie, on: app)
+        try await insertSetup(id: "setup_close")
+        let a = try await insertAssignment(testSetupID: "setup_close", title: "Open", isOpen: true)
+        let id = a.publicID
+
+        try await app.asyncTest(
+            .POST, "/instructor/\(id)/close",
+            beforeRequest: { req in
+                req.headers.add(name: .cookie, value: sessionCookie)
+                try req.content.encode(["_csrf": csrf], as: .urlEncodedForm)
+            },
+            afterResponse: { res in
+                XCTAssertEqual(res.status, .seeOther)
+            })
+
+        let updated = try await APIAssignment.find(a.id, on: app.db)
+        XCTAssertEqual(updated?.isOpen, false)
+    }
+
+    // MARK: - POST /instructor/:id/delete
+
+    func testDeleteAssignmentRemovesRecord() async throws {
+        let cookie = try await loginAsInstructor()
+        let (csrf, sessionCookie) = try await csrfFields(for: "/instructor", cookie: cookie, on: app)
+        try await insertSetup(id: "setup_del")
+        let a = try await insertAssignment(testSetupID: "setup_del", title: "To Remove", isOpen: false)
+        let id = a.publicID
+
+        try await app.asyncTest(
+            .POST, "/instructor/\(id)/delete",
+            beforeRequest: { req in
+                req.headers.add(name: .cookie, value: sessionCookie)
+                try req.content.encode(["_csrf": csrf], as: .urlEncodedForm)
+            },
+            afterResponse: { res in
+                XCTAssertEqual(res.status, .seeOther)
+            })
+
+        let gone = try await APIAssignment.find(a.id, on: app.db)
+        XCTAssertNil(gone)
+    }
+
+    func testDeleteNonexistentAssignmentReturnsNotFound() async throws {
+        let cookie = try await loginAsInstructor()
+        let (csrf, sessionCookie) = try await csrfFields(for: "/instructor", cookie: cookie, on: app)
+        let fakeID = "zzzzzz"
+
+        try await app.asyncTest(
+            .POST, "/instructor/\(fakeID)/delete",
+            beforeRequest: { req in
+                req.headers.add(name: .cookie, value: sessionCookie)
+                try req.content.encode(["_csrf": csrf], as: .urlEncodedForm)
+            },
+            afterResponse: { res in
+                XCTAssertEqual(res.status, .notFound)
+            })
+    }
+
+    // MARK: - POST /instructor/:id/open — nonexistent
+
+    func testOpenNonexistentAssignmentReturnsNotFound() async throws {
+        let cookie = try await loginAsInstructor()
+        let (csrf, sessionCookie) = try await csrfFields(for: "/instructor", cookie: cookie, on: app)
+        let fakeID = "zzzzzz"
+
+        try await app.asyncTest(
+            .POST, "/instructor/\(fakeID)/open",
+            beforeRequest: { req in
+                req.headers.add(name: .cookie, value: sessionCookie)
+                try req.content.encode(["_csrf": csrf], as: .urlEncodedForm)
+            },
+            afterResponse: { res in
+                XCTAssertEqual(res.status, .notFound)
+            })
+    }
+}
